AI and UI Complete

+++ Battle Pets/Assets/Scripts/Ability.cs
@@ -8,10 +8,10 @@ public class Ability
     public ElementType ElementType;
     public AttackType AttackType;
     public int BaseDamage;
-    public float RechargeTime;
+    public int RechargeTime;
 
     // Use this for initialization
-    public Ability(string AbilName, ElementType EltType, AttackType AttType, int BaseDam, float Recharge)
+    public Ability(string AbilName, ElementType EltType, AttackType AttType, int BaseDam, int Recharge)
     {
         Name = AbilName;
         ElementType = EltType;
